backend/pyproject.toml
[tool.poetry]
name = "sports-guide-backend"
version = "0.1.0"
description = ""
authors = ["Sherif Ragimov <ragimov700@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.8"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
flake8 = "^6.1.0"
isort = "^5.13.2"
factory-boy = "^3.3.0"
pyyaml = "^6.0.1"
uritemplate = "^4.1.1"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.1"
django-debug-toolbar = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79
known_first_party = ["my_module", "another_local_module"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
force_sort_within_sections = true
include_trailing_comma = true
multi_line_output = 3
use_parentheses = true
skip = ["*/migrations/*"]